+++ NEWS.md
@@ -1,3 +1,9 @@
+# performance 0.4.8
+
+## Changes to functions
+
+`icc()` now also computes a "classical" ICC for `brmsfit` models. The former way of calculating an "ICC" for `brmsfit` models is now available as new function called `variance_decomposition()`.
+
 # performance 0.4.7
 
 ## General